Acomode Arquitectura
Clases acomodadas en la arquitectura de capas, crear clases de acuerdo a
su paquete
