+++ src/CapaVista/VistasAdminCliente/cliente.java
@@ -1,0 +1,35 @@
+/*
+    BD = Base de Datos
+ */
+package cliente;
+
+/**
+ *
+ * @author Luis
+ */
+public class cliente {
+    private int ID;
+    private String Nombre;
+    private String Telefono;
+    private String Correo;
+    
+    
+    public cliente(String Nombre, String Telefono, String Correo){
+        this.Nombre = Nombre;
+        this.Telefono = Telefono;
+        this.Correo = Correo;
+    }
+    
+    public void agregarBD(){
+        
+    }
+    
+    public void obtenerBD(){
+        
+    }
+    
+    public void eliminarBD(){
+        
+    }
+    
+}
