--- src/mx/uady/accsys/model/SampleModel.java
@@ -1,14 +1,0 @@
-package mx.uady.accsys.model;
-
-/**
- * Created by cesar on 03/09/16.
- */
-public class SampleModel {
-
-    public SampleModel(){
-
-
-
-    }
-
-}
